ops: use uint64 for DocsAndBytes bytes
32 bit integers overflow during moderately large transactions, resulting in incorrect stats being
reported when the number of bytes in a single transaction exceeds the maximum a 32 bit integer can
represent. Using a 64 bit integer allows for much larger byte counts.

+++ crates/derive/src/lib.rs
@@ -31,7 +31,7 @@ trait StatsAccumulator: Default {
 pub struct DocCounter(DocsAndBytes);
 
 impl DocCounter {
-    pub fn new(docs: u32, bytes: u32) -> DocCounter {
+    pub fn new(docs: u32, bytes: u64) -> DocCounter {
         DocCounter(DocsAndBytes { docs, bytes })
     }
 
@@ -42,9 +42,9 @@ impl DocCounter {
 
     /// Add a single document of the given size, incrementing `docs` by 1 and `bytes` by
     /// `doc_byte_len`.
-    pub fn increment(&mut self, doc_byte_len: u32) {
+    pub fn increment(&mut self, doc_byte_len: usize) {
         self.0.docs += 1;
-        self.0.bytes += doc_byte_len;
+        self.0.bytes += doc_byte_len as u64;
     }
 }
